refactor: move createTokens and updateUserRefreshToken to shared utils
