+++ server/src/features/auth/utils/index.ts
@@ -1,3 +1,5 @@
 export * from "./create-jwt.util";
 export * from "./create-payload.util";
+export * from "./create-tokens";
+export * from "./update-user-refresh-token";
 export * from "./verify-password.util";
